Added TextStream layer. Also added the ability for PV to exit early but gracefully.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@7019 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/columns/HyPerCol.hpp
@@ -48,8 +48,8 @@ public:
    int run()  {return run(numSteps);}
    int run(long int nTimeSteps);
 
-   double advanceTime(double time);
-   int   exitRunLoop(bool exitOnFinish);
+   int advanceTime(double time);
+   int exitRunLoop(bool exitOnFinish);
 
    int loadState();
    int columnId();
